Fix format_data()

+++ main.py
@@ -25,5 +25,5 @@ netD = Discriminator(0).to("cpu")
 # print(netD(torch.tensor(test_images[0]).long(), torch.tensor(test_labels[0]).int()))
 # plotImage(netG(z, torch.tensor(test_labels[0]).int()).detach().numpy())
 
-print(netD(test_images[0], test_labels[0]))
+print(netD(test_images[0], test_labels[0].int()))
 plotImage(netG(z, test_labels[0]).detach().numpy())
